Another configuration overhaul Config is now a lot smarter
It is located in dummy.__init__ and is an instance of the Config class to be able to do fancy stuff.
It is now forbidden to import from dummy in your configuration to prevent circular dependencies.
The new configuration style requires runtime imports of collector and statistic classes,
which has been implemented.