Fixed viewer using the new setup
Packaging/modules are A LOT better organised. The now belong to the package that needs them.
That means a lot of modules have been decoupled. Which will do good in SIG tests.

Still alot of testing will need to be done to check if all results are saved and loaded/showed
correctly.